split type errors into module
