+++ teal/variables.lua
@@ -41,4 +41,10 @@ local variables = { Variable = {}, Scope = {} }
 
 
 
+
+
+function variables.has_var_been_used(var)
+   return var.has_been_read_from or var.has_been_written_to
+end
+
 return variables
